refactor(builder): Add authentication check to all protected pages
