+++ apps/builder/pages/builder/[id]/index.tsx
@@ -7,6 +7,12 @@ import { SideMenu } from "../../../features/Builder/SideMenu";
 import { Layout } from "../../../components";
 import { LoadingSpinner, Stack } from "@open-decision/design-system";
 import { BuilderLayout } from "../../../features/Builder/components/BuilderLayout";
+import { GetServerSideProps } from "next";
+import { checkAuthentication } from "../../../features/Auth/checkAuthentication";
+
+export const getServerSideProps: GetServerSideProps = async function (context) {
+  return checkAuthentication(context);
+};
 
 export default function BuilderPage() {
   return (
