+++ apps/builder/pages/index.tsx
@@ -6,48 +6,13 @@ import * as React from "react";
 import { ErrorCard } from "../components/Error/ErrorCard";
 import { getDashboardLayout } from "../features/Dashboard/DashboardLayout";
 import { GetServerSideProps } from "next";
-import { client } from "@open-decision/api-client";
-import { ODError } from "@open-decision/type-classes";
 import { AuthProvider } from "../features/Auth/useAuth";
 import { TLoginOutput } from "@open-decision/auth-api-specification";
 import { TGetTreesOutput } from "@open-decision/tree-api-specification";
+import { checkAuthentication } from "../features/Auth/checkAuthentication";
 
-export const getServerSideProps: GetServerSideProps = async function ({
-  req,
-  res,
-  query,
-}) {
-  const OD = client({ urlPrefix: `${process.env.OD_API_ENDPOINT}/v1` });
-
-  try {
-    const { data: authData, response } = await OD.auth.refreshToken(
-      {},
-      {
-        headers: {
-          Cookie: req.headers.cookie,
-        },
-      }
-    );
-
-    if (authData instanceof ODError) throw response;
-
-    const authenticatedOD = client({
-      token: authData.access.token,
-      urlPrefix: `${process.env.OD_API_ENDPOINT}/v1`,
-    });
-    const { data: trees } = await authenticatedOD.trees.getCollection({});
-
-    const setCookieHeader = response.headers.get("set-cookie");
-    setCookieHeader ? res.setHeader("set-cookie", setCookieHeader) : null;
-    return { props: { ...authData, trees } };
-  } catch (error) {
-    return {
-      redirect: {
-        destination: `/auth/login?from=/`,
-        permanent: false,
-      },
-    };
-  }
+export const getServerSideProps: GetServerSideProps = async function (context) {
+  return checkAuthentication(context);
 };
 
 type PageProps = TLoginOutput & { trees: TGetTreesOutput };
